# openapi/apexrouter-v1.yaml
openapi: 3.1.0
info:
  title: ApexRouter-RS
  version: 0.1.0
  summary: One local router in front of every OpenAI-compatible endpoint you can reach.
  description: >-
    The machine-readable form of ARCHITECTURE.md section 6. Two listeners: the proxy plane
    (the OpenAI/Anthropic drop-in surface) and the control plane (rig, fit, routes, endpoints,
    providers, vast, hf, jobs). Every operation carries an `x-listener` extension naming which
    listener serves it; `/health` is the only path served by both. Prose lives in docs/API.md;
    routing semantics live in docs/ROUTING.md. This file is diffed against the axum route table
    by crates/apexrouter-server/tests/openapi_routes.rs on every CI run.
  license:
    name: MIT OR Apache-2.0
    identifier: MIT OR Apache-2.0
  contact:
    name: ApexRouter-RS
    url: https://github.com/buckster123/ApexRouter-RS
servers:
  - url: http://127.0.0.1:8888
    description: 'proxy listener. `http://127.0.0.1:8888/v1` is equally valid: a doubled leading /v1 is collapsed.'
  - url: http://127.0.0.1:2739
    description: control listener.
...
security:
  - bearerAuth: []
  - tokenHeader: []
  - tokenQuery: []
paths:
...
  /v1/chat/completions:
    post:
      operationId: chatCompletions
      x-listener: [proxy]
      tags: [proxy]
      summary: Routed chat completion, streaming or buffered.
      description: >-
        Relayed byte-for-byte unless the alias differs from the upstream model id, in which case
        exactly one JSON key is rewritten. Never re-framed on a stream.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ChatCompletionRequest' }
      responses:
        '200':
          description: The upstream body, verbatim. `text/event-stream` when the upstream streamed.
          headers:
            X-ApexRouter-Route: { $ref: '#/components/headers/XApexRouterRoute' }
            X-ApexRouter-Backend: { $ref: '#/components/headers/XApexRouterBackend' }
            X-ApexRouter-Attempts: { $ref: '#/components/headers/XApexRouterAttempts' }
            X-ApexRouter-Fallback: { $ref: '#/components/headers/XApexRouterFallback' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ChatCompletionResponse' }
            text/event-stream:
              schema: { type: string }
        '404': { $ref: '#/components/responses/OpenAiError' }
        '413': { $ref: '#/components/responses/OpenAiError' }
        '502': { $ref: '#/components/responses/OpenAiError' }
        '503': { $ref: '#/components/responses/OpenAiError' }
        '504': { $ref: '#/components/responses/OpenAiError' }
        '508': { $ref: '#/components/responses/OpenAiError' }
  /v1/completions:
    post:
      operationId: completions
      x-listener: [proxy]
      tags: [proxy]
      summary: Routed legacy text completion.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CompletionRequest' }
      responses:
        '200':
          description: The upstream body, verbatim.
          content:
            application/json: { schema: { type: object, additionalProperties: true } }
            text/event-stream: { schema: { type: string } }
        '404': { $ref: '#/components/responses/OpenAiError' }
        '503': { $ref: '#/components/responses/OpenAiError' }
...
  /v1/routes/{alias}/test:
    post:
      operationId: testRoute
      x-listener: [control]
      tags: [routes]
      summary: A 20-token probe through the resolved route, using the resolved model id.
      parameters: [{ $ref: '#/components/parameters/Alias' }]
      responses:
        '200':
          description: One probe result.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SmokeProbe' }
        '404': { $ref: '#/components/responses/ControlError' }
...
  /v1/compare:
    post:
      operationId: compare
      x-listener: [control]
      tags: [observability]
      summary: The same prompt against N aliases, at the same time.
      description: Blocking by default, because the answer IS the comparison.
      parameters: [{ $ref: '#/components/parameters/NoWait' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CompareRequest' }
      responses:
        '200':
          description: One row per alias.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/CompareRow' }
        '202':
          description: Accepted; the job's result is the same array.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/JobRecord' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: 'Authorization: Bearer <token>. Scopes read|write|admin are derived from (path, method).'
    tokenHeader:
      type: apiKey
      in: header
      name: X-ApexRouter-Token
      description: The same token, as a header.
    tokenQuery:
      type: apiKey
      in: query
      name: token
      description: The same token, as a query parameter. Never recorded in a tracing span.
  parameters:
...
    NoWait:
      name: no_wait
      in: query
      required: false
      description: Return a JobRecord immediately instead of waiting.
      schema: { type: boolean }
...
    Alias:
      name: alias
      in: path
      required: true
      description: A route alias. Same charset as a BackendId, with '/' banned.
      schema: { $ref: '#/components/schemas/Alias' }
...
  headers:
    XApexRouterRoute:
      description: '<alias-or-"-">|<reason>, e.g. `auto|alias`.'
      schema: { type: string }
    XApexRouterBackend:
      description: Which BackendId answered.
      schema: { type: string }
    XApexRouterAttempts:
      description: How many candidates were tried.
      schema: { type: string }
    XApexRouterFallback:
      description: true when the answer came from a non-first candidate.
      schema: { type: string }
...
  responses:
    OpenAiError:
      description: An OpenAI-shaped error. The dialect always matches the dialect the client spoke.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/OpenAiErrorEnvelope' }
...
    ControlError:
      description: A control-plane ErrorEnvelope. Clients branch on error.kind, never on prose.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  schemas:
...
    BackendId:
      type: string
      pattern: '^[a-z0-9][a-z0-9._-]{0,63}$'
      description: A validated slug. An empty or malformed id is not constructible.
    Alias:
      type: string
      pattern: '^[a-z0-9][a-z0-9._-]{0,63}$'
      description: The string a client puts in "model".
...
    TokenCount:
      type: object
      required: [kind, n]
      properties:
        kind: { type: string, enum: [reported, estimated] }
        n: { type: integer, format: int64 }

    # ---- errors -----------------------------------------------------------------------
    OpenAiErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [message, type]
          properties:
            message: { type: string }
            type:
              type: string
              enum: [model_not_found, upstream_unavailable, upstream_timeout, no_healthy_backend,
                     server_overloaded, request_too_large, loop_detected, provider_not_configured,
                     starting, redacted_endpoint, protocol_not_supported]
            code: { type: [string, 'null'] }
            param: { type: [string, 'null'] }
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error: { $ref: '#/components/schemas/ErrorBody' }
    ErrorBody:
      type: object
      required: [kind, message]
      properties:
        kind:
          type: string
          description: 'The discriminator, so no exit codes had to be invented.'
          examples: [not_found, invalid_request, conflict, port_in_use, unauthorized, forbidden,
                     upstream_unavailable, approval_required, internal]
        message: { type: string }
        param: { type: [string, 'null'] }
        code: { type: [string, 'null'] }
...
    ChatCompletionRequest:
      type: object
      required: [model, messages]
      properties:
        model:
          type: string
          description: >-
            An alias, "<backend_id>/<model>", an upstream model id, or one of the legacy names
            "", "x", "auto", "default" which all fall through to default_alias.
        messages:
          type: array
          items:
            type: object
            properties:
              role: { type: string, enum: [system, user, assistant, tool] }
              content: {}
        stream:
          type: boolean
          description: Read by a strict top-level-key scanner, not a full JSON parse.
        stream_options:
          type: object
          properties:
            include_usage: { type: boolean }
        max_tokens: { type: integer, format: int64 }
        temperature: { type: number }
        tools: { type: array, items: { type: object, additionalProperties: true } }
      additionalProperties: true
    ChatCompletionResponse:
      type: object
      description: The upstream body, verbatim. `model` is the UPSTREAM id, not the alias.
      properties:
        id: { type: string }
        object: { type: string }
        created: { type: integer, format: int64 }
        model: { type: string }
        choices:
          type: array
          items:
            type: object
            properties:
              index: { type: integer, format: int64 }
              message: { type: object, additionalProperties: true }
              finish_reason:
                type: [string, 'null']
                enum: [stop, length, tool_calls, content_filter, null]
        usage: { type: object, additionalProperties: true }
      additionalProperties: true
    CompletionRequest:
      type: object
      required: [model, prompt]
      properties:
        model: { type: string }
        prompt: {}
        max_tokens: { type: integer, format: int64 }
        stream: { type: boolean }
      additionalProperties: true
...
    JobRecord:
      type: object
      required: [id, kind, state, started_unix]
      properties:
        id: { type: string }
        kind: { type: string, examples: [hf.download, vast.rent, endpoint.start, compare, swap] }
        state:
          type: string
          enum: [pending, running, succeeded, failed, cancelled]
          description: Never Pending forever; even a JoinError from a panic flips it to failed.
        pct: { type: [number, 'null'] }
        message: { type: [string, 'null'] }
        started_unix: { type: integer, format: int64 }
        finished_unix: { type: [integer, 'null'], format: int64 }
        result: { description: 'The protocol type the operation produced, as JSON.' }
        error: { type: [string, 'null'] }

    # ---- control plane: rig and fit -----------------------------------------------------
...
    SmokeProbe:
      type: object
      properties:
        name: { type: string, examples: [smoke.models, smoke.warmup, smoke.tools, smoke.throughput] }
        ok: { type: boolean }
        ms: { type: integer, format: int64 }
        detail: { type: string }
        ttft_ms: { type: [integer, 'null'], format: int64 }
        tok_per_s: { type: [number, 'null'] }
        tokens: { type: [integer, 'null'], format: int64 }

    # ---- control plane: endpoints -------------------------------------------------------
...
    CompareRequest:
      type: object
      required: [prompt]
      properties:
        aliases: { type: array, items: { type: string } }
        prompt: { type: string }
        max_tokens: { type: [integer, 'null'], format: int64 }
    CompareRow:
      type: object
      properties:
        alias: { $ref: '#/components/schemas/Alias' }
        backend: { oneOf: [{ $ref: '#/components/schemas/BackendId' }, { type: 'null' }] }
        model: { type: string }
        ok: { type: boolean }
        ms: { type: integer, format: int64 }
        ttft_ms: { type: [integer, 'null'], format: int64 }
        tok_per_s: { type: [number, 'null'] }
        prompt_tokens:
          description: 'The REAL number from the response, never word_count × 1.3.'
          oneOf: [{ $ref: '#/components/schemas/TokenCount' }, { type: 'null' }]
        text: { type: [string, 'null'] }
      additionalProperties: true
...